server/requirements.txt
Django>=3.0.1,<3.1.0
flake8>=3.6.0,<3.7.0
psycopg2-binary>=2.8.3,<2.9.0
celery>=4.2.1,<4.3.0
redis>=2.10.6,<2.11.0
kombu>=4.3.0,<4.4.0
django-cors-headers>=3.2.1,<3.3.0
djangorestframework>=3.11.0,<3.12.0
Pillow>=7.0.0,<7.1.0
gunicorn>=20.0.4,<20.1.0
djangorestframework_simplejwt>=4.4.0,<4.5.0
